[스프링 - 협업 미션] 인비(김태희) 미션 제출합니다. (#2)
* refactor: DTO 기본생성자 추가

* fix: 외래키 참조 구문 테이블명 대소문자 구분 적용

* chore: Jdbc Mysql 드라이버 의존성 추가

* config: 설정 파일 프로파일에 따라 분리

* config: profile active 설정 추가

* fix: DB 테이블명 대소문자 구분 적용

* fix: MemberDao 생성자 테이블명 대소문자 구분 적용

* config: DataLoader profile 설정 변경

* config: Web CORS 허용 설정 추가

* scripts: 배포 shell script 추가

* docs: README 업데이트

* feat: 거리에 따른 추가 요금 계산 기능 추가

* feat: 경로상 노선들의 추가 요금에 따라 총 요금을 계산하는 기능 추가

* refactor: Annotation 순서 수정

* feat: 연령에 따른 요금 할인 기능 추가

* feat: LINE 테이블에 extra_fare 컬럼 추가

* feat: Line 생성 요청에 노선 추가요금 속성 추가 및 함수이름 수정

* feat: 비로그인 상태에서 경로 조회시 총 요금 응답하도록 기능 구현

- 경로 길이에 따른 추가요금 정책 추가
- 노선 추가요금에 따른 정책 추가

* refactor: 함수 이름 변경

* feat: 회원 로그인 시 나이에 따른 차등 요금을 계산해 경로 조회 응답에 포함

* refactor: 초기 노선 데이터에 노선 추가 요금 추가

* style: 자동 포맷팅 적용

* feat: 초기 데이터 insert 조건 추가

* refactor: FareCalculatorTest 패키지 경로 이동

* refactor: test @DisplayName 수정

* feat: GET "/api/lines" API 추가

기존 GET /lines API와 URI만 다를 뿐, 기능 및 응답결과 동일

* style: 자동 포맷팅 적용

* refactor: FareCalculator 패키지 경로 이동

* refactor: 매직넘버 상수화 및 메소드 분리

* refactor: 메소드 순서 정리

* refactor: 메소드 분리 및 라인 정리, 변수이름 수정

* refactor: @DisplayName 수정

* refactor: 테스트 설명 기존 주석 원래 상태로 수정

* docs: README 업데이트

* refactor: null 대신 Optional로 반환하도록 수정

* refactor: 로그인 여부를 LoginMember 객체에 메시지를 던져 확인하도록 수정

* refactor: 새로운 Line이 생성됐을 때, sections까지 초기화 되도록 수정

* refactor: 경로 거리에 따른 추가요금 부과 다형성 적용

* refactor: 패키지 구조 정리

* refactor: 클래스 이름 수정

* refactor: 경로 조회 시, 나이에 따른 할인 적용 다형성 적용

* refactor: 노선 추가 요금 계산 클래스 및 변수 이름 수정

* feat: Request DTO validation 추가

* gradle: rest-docs 의존성 추가

* test: rest-docs 설정

* docs: 회원 인증 API 문서 작성

* refactor: 테스트 파일 패키지 경로 이동

* refactor: 중복 함수 삭제

* test: 유저 관리 rest-docs 설정

* docs: 유저 관리 rest-docs 추가

* test: rest-docs 생성 폴더 이름 수정

* test: 노선 관련 rest-docs 생성 설정 추가

* docs: 노선 관련 rest-docs 추가

* test: 경로 조회 rest-docs 설정 추가

* docs: 경로 조회 관련 rest-docs 추가

* test: 구간 관련 rest-docs 설정 추가

* test: 역 관련 rest-docs 설정 추가

* docs: 역, 구간 관련 rest-docs 추가

* refactor: Delete 할 Station이 없을 때 BadRequest(400)로 응답하도록 수정

* refactor: DTO @Pattern 삭제

* feat: 역 이름 유효성 검사 추가

* fix: @DisplayName 수정

* feat: 노선 이름 유효성 검사 추가

* refactor: Section 테스트 패키지 경로 이동

* feat: 노선 조회 시 구간 정보도 함께 응답하도록 수정

* setting: REST DOCS 설정

* fix: 노선 응답 구간 순서에 맞게 정렬

* refactor: 구간 등록, 삭제 예외 응답코드 및 메세지 수정

* refactor: 사용하지 않는 import 삭제

* docs: README 업데이트

* feat: 노선 생성 응답에 역들과 구간들 정보 포함하도록 수정

* docs: README 업데이트

* test: 정상응답, 예외응답 코드 분리

* feat: 회원가입, 로그인, 내 정보 조회 기능 프론트 요구 예외사항 추가

* feat: 역 관리 기능 프론트 요구 예외사항 추가

* feat: 노선 관리 기능 프론트 요구 예외사항 추가

* feat: 컨트롤러 메서드 500에러 핸들링 추가

* docs: 인텐트 수정

* style: 코드 자동 포맷팅 적용

* refactor: 인증, 멤버 관련 테스트 리팩토링

* refactor: 역 관련 기능 테스트 리팩토링

* refactor: 테스트 클래스에 @DisplayName 추가

* refactor: LineAcceptanceTest 리팩토링

* refactor: 테스트 사용하지 않는 함수 삭제

* refactor: SectionAcceptanceTest 리팩토링

* fix: PathAcceptanceTest 주석 제거

* refactor: Controller 메소드 try~catch 문 삭제

* refactor: Controller 마다의 CORS 허용 설정 별도 설정파일로 분리

* refactor: 변수 final 추가

* refactor: Controller 클래스 레벨의 @RequestMapping 적용

* refactor: Controller 메소드 순서 변경

* refactor: Controller 들의 SQLException 핸들링 AdviceController로 이동

* refactor: Controller들의 ResponseEntity 제네릭 타입 설정

* refactor: 지역변수 제거

* refactor: SectionService 분리, LineService에서 SectionService를 주입받도록 수정

* refactor: Line id 비교 로직 객체 내부로 이동

* refactor: PathService 인스턴스 변수 final 추가

* refactor: 요금 할인 정책 if문 제거

Enum, 함수형 인터페이스, 다형성 이용

* refactor: PathAcceptanceTest 리팩토링

* feat: 예측하지 못한 에러에 대한 @ExceptionHandler 메소드 @ControllerAdvice 클래스 내에 추가

* refactor: SectionAcceptanceTest @DisplayName 수정

* docs: spring rest docs 업데이트

+++ src/main/java/wooteco/subway/station/application/StationService.java
@@ -1,25 +1,34 @@
 package wooteco.subway.station.application;
 
+import java.util.List;
+import java.util.stream.Collectors;
+import org.springframework.dao.DuplicateKeyException;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import wooteco.subway.exception.HttpException;
+import wooteco.subway.line.dao.SectionDao;
 import wooteco.subway.station.dao.StationDao;
 import wooteco.subway.station.domain.Station;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 public class StationService {
-    private StationDao stationDao;
+    private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse saveStation(StationRequest stationRequest) {
-        Station station = stationDao.insert(stationRequest.toStation());
-        return StationResponse.of(station);
+        try {
+            Station station = stationDao.insert(stationRequest.toStation());
+            return StationResponse.of(station);
+        } catch (DuplicateKeyException e) {
+            throw new HttpException(HttpStatus.BAD_REQUEST, "이미 존재하는 역 이름 입니다.");
+        }
     }
 
     public Station findStationById(Long id) {
@@ -30,11 +39,18 @@ public class StationService {
         List<Station> stations = stationDao.findAll();
 
         return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
+            .map(StationResponse::of)
+            .collect(Collectors.toList());
     }
 
     public void deleteStationById(Long id) {
+        Station foundStation = stationDao.findById(id);
+        if (foundStation == null) {
+            throw new HttpException(HttpStatus.BAD_REQUEST, "존재하지 않는 역입니다.");
+        }
+        if (sectionDao.countByStationId(id) > 0) {
+            throw new HttpException(HttpStatus.BAD_REQUEST, "노선에 포함된 역은 삭제할 수 없습니다.");
+        }
         stationDao.deleteById(id);
     }
 }
